Standardized units
All body measurements units in database standardized in metric units
(meters for height, kilograms for weight, and centimeters for everything
else). Not adding the body measurements form in the main page for now
because I could not create an account due to an error (createuser.php
has an error in line 3 - mine at least) or when I tried to create an
account using users_table in phpMyAdmin, I could not login with it
either. And not available until April 5, so sorry. :cry: